requirements-test.txt
# Testing dependencies for the AI coaching platform
# Based on the testing strategy and infrastructure requirements

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # Parallel test execution
pytest-timeout>=2.1.0
pytest-html>=3.2.0  # HTML test reports

# HTTP testing
httpx>=0.24.0
requests>=2.31.0

# Database testing
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Container testing
testcontainers>=3.7.0

# Async testing utilities
asyncio-mqtt>=0.13.0  # For WebSocket testing
websockets>=11.0.0

# Mocking and fixtures
factory-boy>=3.3.0
faker>=19.0.0
responses>=0.23.0

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.16.0  # Load testing

# Security testing
bandit>=1.7.5
safety>=2.3.0
pip-audit>=2.6.0  # Security vulnerability scanning

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Additional flake8 plugins
flake8-docstrings>=1.7.0
flake8-bugbear>=23.7.0
flake8-comprehensions>=3.14.0
flake8-import-order>=0.18.2

# Pre-commit hooks
pre-commit>=3.3.0

# Documentation testing
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# API testing
openapi-spec-validator>=0.6.0
jsonschema>=4.19.0

# Database utilities
alembic>=1.11.0
psycopg2-binary>=2.9.7
asyncpg>=0.28.0

# Redis utilities
redis>=4.6.0
aioredis>=2.0.1

# AI/ML testing utilities
numpy>=1.24.0  # For embedding comparisons
scikit-learn>=1.3.0  # For similarity calculations

# Monitoring and observability
prometheus-client>=0.17.0
structlog>=23.1.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.1.0

# Utilities
click>=8.1.0  # For CLI scripts
rich>=13.5.0  # For beautiful console output
tabulate>=0.9.0  # For table formatting in reports

# Development utilities
ipython>=8.14.0
ipdb>=0.13.0

# Type checking
types-requests>=2.31.0
types-redis>=4.6.0
types-PyYAML>=6.0.0

# Linting and formatting
pylint>=2.17.0
autopep8>=2.0.0

# Coverage reporting
coverage[toml]>=7.3.0

# Test data generation
mimesis>=10.1.0  # Alternative to Faker
hypothesis>=6.82.0  # Property-based testing

# WebSocket testing
pytest-websocket>=0.1.0

# JWT testing
pyjwt>=2.8.0
cryptography>=41.0.0

# File handling for tests
openpyxl>=3.1.0  # Excel files
python-docx>=0.8.11  # Word documents
PyPDF2>=3.0.0  # PDF files

# Time utilities for testing
freezegun>=1.2.0  # Time mocking
python-dateutil>=2.8.0

# Logging testing
pytest-logging>=2016.11.4

# Memory profiling
memory-profiler>=0.61.0
pympler>=0.9

# Network testing
pytest-httpserver>=1.0.8

# Subprocess testing
pytest-subprocess>=1.5.0

# Temporary file utilities
pytest-tmp-files>=0.1.0

# Parameterized testing
pytest-parametrize>=1.1.0

# Retry utilities for flaky tests
pytest-rerunfailures>=12.0

# Test ordering
pytest-order>=1.1.0

# Warnings control
pytest-warnings>=0.3.1

# Environment variable testing
pytest-env>=0.8.2

# Fixtures for common test data
pytest-factoryboy>=2.5.1

# API schema validation
jsonschema-spec>=0.1.4

# Docker testing utilities
docker>=6.1.0

# Process management for tests
psutil>=5.9.0

# Configuration file testing
configparser>=5.3.0
toml>=0.10.2
PyYAML>=6.0.1

# Additional validation utilities
importlib-metadata>=6.8.0  # For better package validation

# Encryption testing
bcrypt>=4.0.0
passlib>=1.7.4

# Rate limiting testing
limits>=3.5.0

# Caching testing
cachetools>=5.3.0

# Serialization testing
msgpack>=1.0.0
orjson>=3.9.0

# Validation testing
cerberus>=1.3.4
marshmallow>=3.20.0

# CLI testing
click-testing>=0.1.0

# Email testing
aiosmtpd>=1.4.4

# File watching for test automation
watchdog>=3.0.0

# Profiling
py-spy>=0.3.14
line-profiler>=4.1.0

# Graph and visualization for test reports
matplotlib>=3.7.0
plotly>=5.15.0

# Statistical analysis for performance tests
scipy>=1.11.0
pandas>=2.0.0

# Network simulation
netifaces>=0.11.0

# System information
platform>=1.0.8

# Concurrent testing
concurrent-futures>=3.1.1

# Queue testing
kombu>=5.3.0

# Workflow testing
celery>=5.3.0

# Health check testing
healthcheck>=1.3.3